move sub module to sep dir
