+++ flipperzero_protobuf/flipperCmd/__main__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-from flipperzero_protobuf.flipperzero_cmd import main
+from .flipperzero_cmd import main
 
 if __name__ == "__main__":
     main()
