+++ flipperzero_protobuf/flipperCmd/flipperzero_cmd.py
@@ -9,9 +9,10 @@ import shlex
 import argparse
 
 
-from .flipperCmd import FlipperCMD
+#from .flipperCmd import FlipperCMD
+from . import FlipperCMD
 # from google.protobuf.json_format import MessageToDict
-from .flipper_base import cmdException    # FlipperProtoBase
+from ..flipper_base import cmdException    # FlipperProtoBase
 # from .flipper_storage import FlipperProtoStorage
 # from .flipper_proto import FlipperProto
 # from .cli_helpers import print_screen, flipper_tree_walk, calc_file_md5
